feat(middleware): added setupPersistedMessage middleware
logs a message after recording is updated
